// client/src/app/models/roles.ts
export enum Roles {
  ROLE_SUPER_ADMIN = 'ROLE_SUPER_ADMIN',
  ROLE_ADMIN = 'ROLE_ADMIN',
  ROLE_STAFF_DATA_MANAGER = 'ROLE_STAFF_DATA_MANAGER',
  ROLE_STAFF_DATA_VIEWER = 'ROLE_STAFF_DATA_VIEWER',
  ROLE_PARTNER_DATA_VIEWER = 'ROLE_PARTNER_DATA_VIEWER',
  ROLE_PARTNER_DATA_MANAGER = 'ROLE_PARTNER_DATA_MANAGER'
}